Add communication mixin and common formatting helpers
